add user's financial score

+++ backend/src/main/java/com/eyes/solstory/domain/financial/controller/FinancialSummaryController.java
@@ -142,14 +142,13 @@ public class FinancialSummaryController {
      * 저축 점수 : 기본 100 + 지난 달 대비 이번 달 저축 증감률 (0 ~ 200)  
      * 지출 점수 : 기본 100 - 지난 달 대비 이번 달 소비 증감률 (0 ~ 200)
      * >> 분석 점수 : (저축 + 지출)/4 > 0 ~ 100점까지 (소수점 두자리 수까지 반환)
+     *  = 기본 50 + (저축 점수 + 지출 점수)/4 
      * @param userNo
      * @return
      */
-    /*
     @GetMapping("/financial-score")
-    public ResponseEntity<Double> getFinancialScore(@RequestParam("userNo") int userNo) {
+    public ResponseEntity<Integer> getFinancialScore(@RequestParam("userNo") int userNo) {
     	return ResponseEntity.ok(summaryAnalyzer.getFinancialScore(userNo));
     }
-    */
     
 }
